requirements.txt
streamlit
openai
python-dotenv
edge_tts
requests